+++ Source/NetOffice/CoreExtensions/CoreFactoryExtensions.cs
@@ -41,12 +41,11 @@ namespace NetOffice
         /// <param name="hostCache">core host cache</param>
         /// <param name="caller">calling instance</param>
         /// <param name="comProxy">new created proxy</param>
-        /// <param name="wantTheDuck">want duck implementation</param>
         /// <param name="throwException">throw exception if no info found or return null</param>
         /// <returns>factory info from corresponding assembly</returns>
         internal static IFactoryInfo GetFactoryInfo(this Core value, 
             Dictionary<Guid, Guid> hostCache, ICOMObject caller,
-            object comProxy, bool wantTheDuck, bool throwException)
+            object comProxy, bool throwException)
         {
             if (value.Assemblies.Count == 0)
                 return null;
@@ -59,7 +58,7 @@ namespace NetOffice
             {             
                 foreach (IFactoryInfo item in value.Assemblies)
                 {                   
-                    if (item.IsDuck != wantTheDuck || item.AssemblyName != caller.InstanceComponentName)
+                    if (item.AssemblyName != caller.InstanceComponentName)
                         continue;
                     foreach (var guid in item.ComponentGuid)
                         if (true == guid.Equals(hostGuid))
@@ -70,8 +69,6 @@ namespace NetOffice
             {
                 foreach (IFactoryInfo item in value.Assemblies)
                 {
-                    if (item.IsDuck != wantTheDuck)
-                        continue;
                     foreach (var guid in item.ComponentGuid)
                         if (true == guid.Equals(hostGuid))
                             return item;
@@ -82,8 +79,6 @@ namespace NetOffice
             // list of known multiple defined types is available in Attributes\DuplicateAttribute.cs file
             foreach (IFactoryInfo item in value.Assemblies)
             {
-                if (item.IsDuck != wantTheDuck)
-                    continue;
                 bool hasComponentID = null != item.ComponentGuid ? item.ComponentGuid.Contains(hostGuid) : false;
                 if (item.Contains(className) && hasComponentID)
                 {